Removed unnecessary static_cast.

+++ CsvTable.cpp
@@ -108,8 +108,7 @@ void FileLines::getPositionsOfSampleLines()
         if (mNumLines == kMinNumLines) {
             // Evaluate number of lines, excluding headers' line, in the file
             assert(mFileStream && mFileStream.tellg() > 0);
-            const auto approxNumLines
-                = mNumLines * (mFileSize - mPosSampleLine.at(1)) / (static_cast<std::size_t>(mFileStream.tellg()) - mPosSampleLine.at(1));
+            const auto approxNumLines = mNumLines * (mFileSize - mPosSampleLine.at(1)) / (mFileStream.tellg() - mPosSampleLine.at(1));
             BOOST_LOG_SEV(gLogger, bltrivial::trace)
                 << "file_size=" << mFileSize << ", approxNumLines=" << approxNumLines << FUNCTION_FILE_LINE;
             assert(approxNumLines > 0);
